Add files via upload
This code calculates the Armstrong sum of a given integer x. The Armstrong sum is the sum of each digit of the number raised to the power of the total number of digits in the number. However, the code does not check whether the input number is an Armstrong number; it only computes the sum.